security. jdbc: add mapping on registration and root

+++ ltale/src/main/java/com/vad/ltale/entity/User.java
@@ -9,7 +9,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_user")
-    private int id;
+    private Long id;
 
     @Column(name = "username")
     private String username;
@@ -40,7 +40,7 @@ public class User {
         this.role = role;
     }
 
-    public User(int id, String username, String email, String password, int enabled, Role role) {
+    public User(long id, String username, String email, String password, int enabled, Role role) {
         this.id = id;
         this.username = username;
         this.email = email;
@@ -51,11 +51,11 @@ public class User {
 
     // define getter/setter
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
